Update virtual-thermostat-with-device.groovy
Changed code for cooling instead of heating, but broke it in the process.  Now the outlet switch turns on when the desired temperature is set below the current temperature, as expected.  When, however, the desired temperature is set above the current temperature, the on command is again sent to the outlet switch, instead of an off command.